+++ internal/logging/ansi.go
@@ -1,0 +1,9 @@
+package logging
+
+import "regexp"
+
+var ansiRE = regexp.MustCompile(`\x1b\[[0-9;]*m`)
+
+func stripANSI(s string) string {
+	return ansiRE.ReplaceAllString(s, "")
+}